airdrop: cleanup code structure
