+++ contracts/airdrop/src/execute/mod.rs
@@ -1,0 +1,12 @@
+mod claim;
+mod close;
+mod fund;
+mod register;
+
+pub use crate::execute::claim::{claim, claim_many};
+pub use crate::execute::close::close;
+pub use crate::execute::fund::fund;
+pub use crate::execute::register::register;
+
+#[cfg(test)]
+mod test;
