+++ contracts/airdrop/src/state.rs
@@ -3,15 +3,31 @@ use cosmwasm_std::{Addr, Uint128};
 use cw_storage_plus::{Item, Map};
 
 #[cw_serde]
-pub struct Airdrop {
-    pub creator: Addr,
-    pub merkle_root: String,
-    pub denom: String,
-    pub total_amount: Uint128,
-    pub total_claimed: Uint128,
-    pub label: Option<String>,
-    pub bearer: bool,
-    pub closed: bool,
+pub enum Airdrop {
+    Open {
+        creator: Addr,
+
+        denom: String,
+        total_amount: Uint128,
+        total_claimed: Uint128,
+        merkle_root: String,
+
+        label: Option<String>,
+        closed: bool,
+    },
+
+    Bearer {
+        creator: Addr,
+        signer: Addr,
+
+        denom: String,
+        total_amount: Uint128,
+        total_claimed: Uint128,
+        merkle_root: String,
+
+        label: Option<String>,
+        closed: bool,
+    },
 }
 
 pub const LATEST_AIRDROP_KEY: &str = "latest_airdrop";
