+++ packages/interface/src/airdrop.rs
@@ -29,61 +29,64 @@ pub enum AirdropIdOptional {
 }
 
 #[cw_serde]
-pub enum ClaimProof {
-    Account(String),
-    ClaimProof(String),
+pub enum RegisterPayload {
+    // Payload for open airdrop
+    Open {
+        // merkle root of airdrop
+        merkle_root: String,
+
+        // denomination of airdrop
+        denom: String,
+
+        // optional: label
+        label: Option<String>,
+    },
+
+    // Payload for bearer airdrop
+    Bearer {
+        merkle_root: String,
+        denom: String,
+        label: Option<String>,
+        signer: Option<String>,
+    },
 }
 
 #[cw_serde]
-pub enum ClaimProofOptional {
-    Account(Option<String>),
-    ClaimProof(String),
-}
+pub enum ClaimPayload {
+    // Payload for open airdrop
+    Open {
+        airdrop: AirdropId,        // airdrop specifier
+        amount: Uint128,           // claim amount
+        account: Option<String>,   // address who claims
+        merkle_proof: Vec<String>, // merkle proof of airdrop
+    },
 
-impl ClaimProofOptional {
-    pub fn account(address: impl Into<String>) -> Self {
-        ClaimProofOptional::Account(Some(address.into()))
-    }
-
-    pub fn claim_proof(proof: impl Into<String>) -> Self {
-        ClaimProofOptional::ClaimProof(proof.into())
-    }
-}
-
-#[cw_serde]
-pub struct ClaimPayload {
-    pub id: AirdropId,
-    pub amount: Uint128,
-    pub claim_proof: ClaimProofOptional,
-    pub merkle_proof: Vec<String>,
+    // Payload for bearer airdrop
+    Bearer {
+        airdrop: AirdropId,        // airdrop specifier
+        amount: Uint128,           // claim amount
+        claim_hash: String,        // salt hash to prevent double spending
+        claim_sign: String,        // signature of signer
+        merkle_proof: Vec<String>, // merkle proof of airdrop
+    },
 }
 
 #[cw_serde]
 pub enum ExecuteMsg {
-    Register {
-        merkle_root: String,
-        denom: String,
-        label: Option<String>,
-        bearer: Option<bool>,
-    },
-
-    Fund {
-        id: AirdropId,
-    },
+    Register(RegisterPayload),
+    Fund(AirdropId),
 
     Claim(ClaimPayload),
     MultiClaim(Vec<ClaimPayload>),
 
-    Close {
-        id: AirdropId,
-    },
+    Close(AirdropId),
 }
 
 #[cw_serde]
 #[derive(QueryResponses)]
 pub enum QueryMsg {
     #[returns(GetAirdropResponse)]
-    GetAirdrop { id: AirdropId },
+    GetAirdrop(AirdropId),
 
     #[returns(ListAirdropsResponse)]
     ListAirdrops {
@@ -97,38 +100,50 @@ pub enum QueryMsg {
 
     #[returns(GetClaimResponse)]
     GetClaim {
-        id: AirdropId,
-        claim_proof: ClaimProof,
+        airdrop: AirdropId,
+        claim_key: String,
     },
 
     #[returns(ListClaimsResponse)]
     ListClaims {
-        id: AirdropId,
+        airdrop: AirdropId,
         start_after: Option<String>,
         limit: Option<u32>,
         order: Option<RangeOrder>,
     },
 
-    #[returns(CheckQualificationResponse)]
-    CheckQualification {
-        id: AirdropId,
-        amount: Uint128,
-        claim_proof: ClaimProof,
-        merkle_proof: Vec<String>,
-    },
+    #[returns(ValidateClaimResponse)]
+    ValidateClaim(ClaimPayload),
 }
 
 #[cw_serde]
-pub struct GetAirdropResponse {
-    pub id: u64,
-    pub creator: String,
-    pub label: Option<String>,
-    pub denom: String,
-    pub total_amount: Uint128,
-    pub total_claimed: Uint128,
-    pub merkle_root: String,
-    pub bearer: bool,
-    pub closed: bool,
+pub enum GetAirdropResponse {
+    Open {
+        id: u64,
+        creator: String,
+
+        denom: String,
+        total_amount: Uint128,
+        total_claimed: Uint128,
+        merkle_root: String,
+
+        label: Option<String>,
+        closed: bool,
+    },
+
+    Bearer {
+        id: u64,
+        creator: String,
+        signer: String,
+
+        denom: String,
+        total_amount: Uint128,
+        total_claimed: Uint128,
+        merkle_root: String,
+
+        label: Option<String>,
+        closed: bool,
+    },
 }
 
 #[cw_serde]
@@ -139,15 +154,16 @@ pub struct LatestAirdropResponse(pub u64);
 
 #[cw_serde]
 pub struct GetClaimResponse {
+    pub id: u64,
     pub amount: Uint128,
-    pub claim_proof: ClaimProof,
+    pub claim_key: String,
 }
 
 #[cw_serde]
 pub struct ListClaimsResponse(pub Vec<GetClaimResponse>);
 
 #[cw_serde]
-pub struct CheckQualificationResponse(pub bool);
+pub struct ValidateClaimResponse(pub bool);
 
 #[cw_serde]
 pub struct MigrateMsg {
